% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqs.R
\name{getGenBankSeqs}
\alias{getGenBankSeqs}
\title{Download sequences and metadata from GenBank by unique identifier}
\usage{
getGenBankSeqs(id)
}
\arguments{
\item{id}{character vector of unique ID(s) for records in database}
}
\value{
a list with two elements: \code{data} contains the metadata for each
sequence; and \code{dna} contains a character vector (same length as number
of rows of \code{data}) where each element is a FASTA-formatted DNA sequence
matched to the metadata by accession.version number

 @examples
 \dontrun{
 foo <- getGenBankSeqs(c('1331395866', '1679378317'))
}
}
\description{
Download sequences and metadata from GenBank by unique identifier
}
