% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Return taxa metadata based on ids}
\usage{
getMetadata(ids)
}
\arguments{
\item{ids}{ID or IDs of species to search for}
}
\value{
Return data frame of species ids, organism, create_date, lat_lon, product, organelle, and sequence.
}
\description{
Return taxa metadata based on ids
}
\examples{
\dontrun{
getMetadata(2029858529)
getMetadata(c(2029858529, 257154225, 257154223, 257154221))
}
}
