% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqIDs.R
\name{getNCBISeqID}
\alias{getNCBISeqID}
\title{Download unique identifiers by taxon for sequences on NCBI}
\usage{
getNCBISeqID(spp)
}
\arguments{
\item{spp}{vector of species names (should already be scrubbed)}
}
\value{
a \code{data.frame} with columns \code{uid} and \code{gid} that
record (respectively) the species ID and the sequence ID from NCBI

 @examples
 \dontrun{
 foo <- getNCBISeqID(c('Drosophila murphyi', 'Drosophila sproati'))
}
}
\description{
Download unique identifiers by taxon for sequences on NCBI
}
