% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonomy.R
\name{getNCBITaxonomy}
\alias{getNCBITaxonomy}
\title{Download taxonomy (with error correcting) from NCBI}
\usage{
getNCBITaxonomy(spp)
}
\arguments{
\item{spp}{character vector of species names}
}
\value{
a \code{data.frame} with columns \code{old_name}, \code{ncbi_name},
\code{uid} that record (respectively) the user-supplied name, the name
matched in NCBI, the species ID from NCBI

 @examples
 \dontrun{
 foo <- getNCBITaxonomy(c('Idiomyia sproati', 'Drosophil murphy', 'no body'))
}
}
\description{
Download taxonomy (with error correcting) from NCBI
}
